Menu: Move menu into widgets folder
